refactor(frontend): extract ApiTokenForm component, reuse for bot tokens
